build(deploy): use stable Yarn version for consistency
Replace the use of actions/setup-node with corepack to ensure the latest stable
version of Yarn is activated for deployment. This change removes the need for the
borales/actions-yarn action and streamlines dependency installation and project
building with Yarn commands.

By aligning with the stable Yarn version through corepack, the deployment process
benefits from the latest Yarn features and security updates, improving the reliability
and security of the deployment workflow.